Move srcSet functionalities to a utilities folder
